refactor: consolidate timetable generation and improve data model constraints
- Remove deprecated SchedulerTestController with extensive mock data
- Simplify TimetableTestController to use TimetableGeneratorService directly
- Add unique constraints to Course.code and Department.name for data integrity
- Add validation pattern for Course.code (format: ABC123)
- Refactor TimetableGeneratorServiceImpl to remove try-catch wrapper and simplify flow
- Update VenueSeeder with specific engineering department codes (PETE, MECH, CIVIL, EIE)
- Remove duplicate Chapel venue entry in VenueSeeder
- Move mock assignment creation logic into TimetableGeneratorServiceImpl for reuse
