+++ src/main/java/com/henry/universitycourseschedular/models/Department.java
@@ -14,6 +14,7 @@ public class Department {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(unique = true)
     private String name;
 
     @Column(unique = true)
